requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
dnspython==2.2.1
mongoengine==0.24.2
pymongo==4.2.0
pytz==2022.2.1
sqlparse==0.4.2
